GH-503 loading state
