+++ SafeUIKit/ReplaceBrowserExtensionUI/Common/FeeCalculation/FeeCalculationLine.swift
@@ -5,12 +5,16 @@
 import Foundation
 import UIKit
 
-public class FeeCalculationLine {
+public class FeeCalculationLine: Equatable {
 
     var lineHeight: Double = 25
 
     func makeView() -> UIView {
         return UIView()
+    }
+
+    public static func ==(lhs: FeeCalculationLine, rhs: FeeCalculationLine) -> Bool {
+        return lhs === rhs
     }
 
 }
