+++ SafeUIKit/ReplaceBrowserExtensionUI/en.lproj/Localizable.strings
@@ -1,3 +1,18 @@
+/* Balance */
+"fee_calculation.balance" = "Balance";
+
+/* Current balance */
+"fee_calculation.current_balance" = "Current balance";
+
+/* Network fee */
+"fee_calculation.fee" = "Network fee";
+
+/* [?] */
+"fee_calculation.fee.info" = "[?]";
+
+/* - ETH */
+"fee_calculation.value.loading_eth" = "- ETH";
+
 /* Loading... */
 "loading_title_text" = "Loading…";
